Don't move deprecated classes to deprecated folder; leave in existing location
