+++ Dnn.CommunityForums/DnnCommunityForums.csproj
@@ -1126,15 +1126,18 @@
     <Compile Include="Controllers\TopicRatingController.cs" />
     <Compile Include="Controllers\UrlController.cs" />
     <Compile Include="Controllers\UserController.cs" />
-    <Compile Include="Deprecated\LikeController.cs" />
-    <Compile Include="Deprecated\LikeInfo.cs" />
-    <Compile Include="Entities\LikeInfo.cs" />
     <Compile Include="Entities\FilterInfo.cs" />
     <Compile Include="Deprecated\Cache.cs" />
     <Compile Include="Deprecated\Utilities.cs" />
-    <Compile Include="Deprecated\Cache.cs" />
-    <Compile Include="Deprecated\Cache.cs" />
+    <Compile Include="Entities\Like.cs" />
+    <Compile Include="Entities\LikeInfo.cs" />
+    <Compile Include="Entities\FilterInfo.cs" />
     <Compile Include="Deprecated\Utilities.cs" />
+    <Compile Include="Entities\Like.cs" />
+    <Compile Include="Entities\LikeInfo.cs" />
+    <Compile Include="Entities\FilterInfo.cs" />
+    <Compile Include="Deprecated\Utilities.cs" />
+    <Compile Include="Deprecated\Cache.cs" />
     <Compile Include="Controllers\LikeController.cs" />
     <Compile Include="class\ParamBuilder.cs" />
     <Compile Include="class\StringExtensions.cs" />
@@ -1240,8 +1243,8 @@
     <Compile Include="components\Data\Topics.cs" />
     <Compile Include="components\Extensions\ConfigUtils.cs" />
     <Compile Include="components\Extensions\ReWriter.cs" />
-    <Compile Include="Deprecated\FilterController.cs" />
-    <Compile Include="Deprecated\FilterInfo.cs" />
+    <Compile Include="components\Filters\FilterController.cs" />
+    <Compile Include="components\Filters\FilterInfo.cs" />
     <Compile Include="components\Forums\ForumCollection.cs" />
     <Compile Include="components\Helpers\Environment.cs" />
     <Compile Include="components\Helpers\JSON.cs" />
