+++ Dnn.CommunityForums/class/ForumController.cs
@@ -21,7 +21,6 @@
 using System;
 using System.Data;
 using System.IO;
-using System.Runtime.CompilerServices;
 using System.Text;
 using System.Xml;
 using DotNetNuke.Modules.ActiveForums.Data;
